requirements/dev.txt
-r common.txt

# ^^^^^^^^^^^^^^^^^^^^^^
# DJANGO DEBUG TOOLBAR
# Also requires: Django
# ^^^^^^^^^^^^^^^^^^^^^^
django-debug-toolbar==1.9.1

# ^^^^^^^^^^^^^^^^^^^^^^
# TESTING
# ^^^^^^^^^^^^^^^^^^^^^^
coverage>=4.4.2

# ^^^^^^^^^^^^^^^^^^^^^^
# DOCUMENTATION
# ^^^^^^^^^^^^^^^^^^^^^^
coreapi==2.3.3

# ^^^^^^^^^^^^^^^^^^^^^^
# MISC
# ^^^^^^^^^^^^^^^^^^^^^^
ipython==6.2.1
isort>=4.2.15
pylint>=1.7.4
